sdk: Move target-specific HTTP code into new submodules
